Fix Task2 switching to integer numbers
As we are counting the number of digits in a number and not its representation but the floating-point numbers do not have that property the solution was switched to Long instead.

+++ src/main/java/edu/hw1/Task2.java
@@ -7,8 +7,8 @@ public class Task2 {
         throw new IllegalStateException();
     }
 
-    public static int countDigits(double number) {
-        double left = number;
+    public static int countDigits(long number) {
+        long left = number;
         int result = 1;
         while (left > BASE) {
             result++;
